+++ esp32/src/graphics/matrix.h
@@ -1,11 +1,10 @@
 #pragma once
 
-#ifdef UNIT_TEST
-#include "../test/mocks/mock_fastled.h"
+#include "hal/types.h"
+
+#ifndef ESP32
 #include <string>
 using String = std::string;
-#else
-#include <FastLED.h>
 #endif
 
 #include "coordinate_transforms.h"